Move tests to plots module where possible. Increase coverage, cleanup.
