+++ plots/src/test/kotlin/org/cryptobiotic/rlauxe/sim/CorlaTask.kt
@@ -1,7 +1,7 @@
 package org.cryptobiotic.rlauxe.sim
 
 import org.cryptobiotic.rlauxe.core.Cvr
-import org.cryptobiotic.rlauxe.sampling.ComparisonWithErrorRates
+import org.cryptobiotic.rlauxe.cobra.ComparisonWithErrorRates
 import org.cryptobiotic.rlauxe.core.RiskTestingFn
 import org.cryptobiotic.rlauxe.sampling.SampleGenerator
 import org.cryptobiotic.rlauxe.corla.Corla
